requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-wtf==1.1.1
wtforms==3.0.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pillow==10.0.1
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0

# Optional (recommended) for FaceNet-based recognition
# Note: install torch according to your platform (CPU vs CUDA) per https://pytorch.org
facenet-pytorch>=2.5.2
torch>=2.0.0
torchvision>=0.15.0
